requirements.txt
langchain
faiss-cpu
gradio
PyPDF2
openai
jinja2 
typing-extentions
tiktoken
python-dotenv